refactor(library/Base64CharacterEncodedByteSequence): moves file out of "/customTypes"
