+++ src/features/TextToImage/models/SDXL/client/conversions/StabilityAI_Client_Image.ts
@@ -1,6 +1,6 @@
 import type * as StabilityAIClient from 'stabilityai-client-typescript/models/components';
 
-import Base64CharacterEncodedByteSequence from '@/library/customTypes/Base64CharacterEncodedByteSequence.ts';
+import Base64CharacterEncodedByteSequence from '@/library/Base64CharacterEncodedByteSequence';
 import ImageURI from '@/library/customTypes/UniformResourceIdentifier/Data/Image/index.ts';
 
 import type {
