+++ src/library/Base64CharacterEncodedByteSequence.ts
