+++ src/library/customTypes/UniformResourceIdentifier/Data/Image/index.ts
@@ -1,4 +1,4 @@
-import type Base64CharacterEncodedByteSequence from '@/library/customTypes/Base64CharacterEncodedByteSequence.ts';
+import type Base64CharacterEncodedByteSequence from '@/library/Base64CharacterEncodedByteSequence';
 
 import type {
   DataURIBinaryEncoding,
